Commenting. Fix for web service bug. Moving more interfaces.
