+++ src/app/add-release/add-release.component.ts
@@ -11,12 +11,18 @@ import {COMMA, ENTER} from '@angular/cdk/keycodes';
 })
 export class AddReleaseComponent implements OnInit {
 
+  // Constants
   readonly separatorKeysCodes: number[] = [ENTER, COMMA];
 
+  // Form for FormControl?/Builder?
   form;
+
+  // Data for select components
   clients = [{name: 'client1'}, {name: 'client2'}, {name: 'client3'}];
   routes = [{route: 'route1'}, {route: 'route2'}, {route: 'route3'}];
   contTypes = [{type: 'cont-type1'}, {type: 'cont-type2'}, {type: 'cont-type3'}];
+
+  // List for the container numbers
   containerNumbers: string[] = ['Cont-1-a', 'Cont-2-a', 'Cont-3-a'];
 
   // Inject the WebService and FormBuilder
@@ -29,6 +35,7 @@ export class AddReleaseComponent implements OnInit {
   ngOnInit() {
   }
 
+  // Adds item to the chip list
   add(event: MatChipInputEvent): void {
     const input = event.input;
     const value = event.value;
@@ -42,6 +49,7 @@ export class AddReleaseComponent implements OnInit {
     }
   }
 
+  // Removes items from chip list
   remove(containerNumber: string): void {
     const index = this.containerNumbers.indexOf(containerNumber);
 
@@ -50,11 +58,13 @@ export class AddReleaseComponent implements OnInit {
     }
   }
 
+  // Returns current time to fill the form with time values
   getTime() {
     let date = new Date();
     return (date.getHours() < 10 ? '0' + date.getHours() : date.getHours()) + ':' + date.getMinutes();
   }
 
+  // Returns current date to fill the form with date values
   getDate() {
     return new Date();
   }
