+++ src/app/edit-release/edit-release.component.html
@@ -1,3 +1,9 @@
-<p>
-  edit-release works!
-</p>
+<table mat-table>
+  <th>Releases</th>
+  <tr>
+    <td>First release</td>
+    <td>
+      <button mat-button>edit</button>
+    </td>
+  </tr>
+</table>
