+++ src/app/home/home.component.ts
@@ -8,11 +8,11 @@ import {WebService} from '../web.service';
 })
 export class HomeComponent implements OnInit {
 
-  // Injects the webservice
+  // Injects the WebService
   constructor(public webService: WebService) {
   }
 
-  // Gets the messages from the webservice
+  // Gets the messages from the WebService
   ngOnInit() {
     this.webService.spinUpAPI();
   }
