+++ src/app/interfaces.ts
@@ -1,18 +1,22 @@
+// Interface to describe the days rounds (trucks) containing all the trucks and their associated rounds
 export interface Trucks {
 	rounds: TruckRounds[]
 }
 
+// Interface for a Trucks rounds for a given day
 export interface TruckRounds {
 	id: string;
 	dayRounds: Rounds[];
 	nightRounds: Rounds[];
 }
 
+// Interface to describe a Round that a truck owns
 export interface Rounds {
 	roundNumber: number;
 	slots: Slot[];
 }
 
+// Interface that describes a slot in a truck's round
 export interface Slot {
 	release: Release;
 }
@@ -24,3 +28,33 @@ export interface Release {
 	qty: number;
 	colour: string;
 }
+
+// Initial draft interface for a Full Releases stored by the API
+export interface FullRelease {
+  release: string;
+  qtyForty: number;
+  qtyTwenty: number;
+  colour: string;
+}
+
+// Interface to describe how a warning popup should be constructed
+// message: the string to be displayed in the dialog as a header
+// body:    string array for the message to be displayed in the body. 0=Main Body 1=Unused
+//          (if 0 & 1) 0=Component Defined 1 1= Component Defined 2
+// options: boolean array specifying which buttons to enable in the popup 0=OK 1=Cancel 2=Component Defined 1
+//          3=Component Defined 2 4=Unused
+// result:  string array of data to be returned depending on button press 0=Click outside dialog 1=OK 2=Cancel 3=Component Defined 1
+//          4=Component Defined 2 5=Unused
+export interface Warning {
+  message: string;
+  body: string[];
+  options: boolean[];   // Defines what buttons should be available
+  result: string[];     // Defines the form of response from the dialog
+}
+
+// Interface for a representation of a day
+export interface Day {
+  day: string;
+  disabled: boolean;
+}
+
