+++ src/app/planning/planning.component.ts
@@ -8,22 +8,21 @@ import {WebService} from '../web.service';
 })
 export class PlanningComponent implements OnInit {
 
-  // Injects the webservice
+  // Injects the WebService
   constructor(private webService: WebService) {}
 
-  // Updates the component and children to the current day.
   ngOnInit() {
-    // this.changeDay(this.getCurrentDay());
   }
 
   // Retrieves the current day in a format the API will understand
+  // Unused?
   private getCurrentDay() {
     const date = new Date();
 
     return date.getDate() + '-' + (date.getMonth() + 1) + '-' + date.getFullYear();
   }
 
-  // Causes the webservice to retrieve the specified day
+  // Causes the WebService to retrieve the specified day
   changeDay(day) {
     this.webService.setCurrentDay(day);
     this.webService.getReleases(day);
