+++ src/app/release-information/release-information.component.ts
@@ -1,6 +1,7 @@
 import {Component, Inject, OnInit} from '@angular/core';
-import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material';
+import {MAT_DIALOG_DATA} from '@angular/material';
 import {WebService} from '../web.service';
+import {FullRelease} from '../interfaces';
 
 @Component({
   selector: 'app-release-information',
@@ -13,7 +14,8 @@ export class ReleaseInformationComponent implements OnInit {
   fullRelease: FullRelease;
 
   // Injects the webservice and other data from the parent component
-  constructor(/* public dialogRef: MatDialogRef<ReleaseInformationComponent>, */
+  constructor(
+              /* public dialogRef: MatDialogRef<ReleaseInformationComponent>, */
               @Inject(MAT_DIALOG_DATA) public data: string,
               private webService: WebService) {
   }
@@ -27,13 +29,5 @@ export class ReleaseInformationComponent implements OnInit {
       this.fullRelease = this.webService.fullReleaseStore;
     });
   }
-
 }
 
-// Initial draft interface for a Full Releases information
-export interface FullRelease {
-  release: string;
-  qtyForty: number;
-  qtyTwenty: number;
-  colour: string;
-}
