+++ src/app/warning-popup/warning-popup.component.ts
@@ -1,5 +1,6 @@
 import {Component, Inject, OnInit} from '@angular/core';
 import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material';
+import {Warning} from '../interfaces';
 
 @Component({
   selector: 'app-warning-popup',
@@ -53,17 +54,4 @@ export class WarningPopupComponent implements OnInit {
   }
 }
 
-// Interface for the data input type
-// message: the string to be displayed in the dialog as a header
-// body:    string array for the message to be displayed in the body. 0=Main Body 1=Unused
-//          (if 0 & 1) 0=Component Defined 1 1= Component Defined 2
-// options: boolean array specifying which buttons to enable in the popup 0=OK 1=Cancel 2=Component Defined 1
-//          3=Component Defined 2 4=Unused
-// result:  string array of data to be returned depending on button press 0=Click outside dialog 1=OK 2=Cancel 3=Component Defined 1
-//          4=Component Defined 2 5=Unused
-export interface Warning {
-  message: string;
-  body: string[];
-  options: boolean[];   // Defines what buttons should be available
-  result: string[];     // Defines the form of response from the dialog
-}
+
